Restore generated solidity files clobbered by local build

+++ linera-bridge/src/solidity/WrappedFungibleTypes.sol
@@ -3,16 +3,11 @@ pragma solidity ^0.8.0;
 import "BridgeTypes.sol";
 
 library WrappedFungibleTypes {
-
     struct ApplicationId {
         BridgeTypes.CryptoHash application_description_hash;
     }
 
-    function bcs_serialize_ApplicationId(ApplicationId memory input)
-        internal
-        pure
-        returns (bytes memory)
-    {
+    function bcs_serialize_ApplicationId(ApplicationId memory input) internal pure returns (bytes memory) {
         return BridgeTypes.bcs_serialize_CryptoHash(input.application_description_hash);
     }
 
@@ -27,11 +22,7 @@ library WrappedFungibleTypes {
         return (new_pos, ApplicationId(application_description_hash));
     }
 
-    function bcs_deserialize_ApplicationId(bytes memory input)
-        internal
-        pure
-        returns (ApplicationId memory)
-    {
+    function bcs_deserialize_ApplicationId(bytes memory input) internal pure returns (ApplicationId memory) {
         uint256 new_pos;
         ApplicationId memory value;
         (new_pos, value) = bcs_deserialize_offset_ApplicationId(0, input);
@@ -44,11 +35,7 @@ library WrappedFungibleTypes {
         uint128 amount;
     }
 
-    function bcs_serialize_BurnEvent(BurnEvent memory input)
-        internal
-        pure
-        returns (bytes memory)
-    {
+    function bcs_serialize_BurnEvent(BurnEvent memory input) internal pure returns (bytes memory) {
         bytes memory result = bcs_serialize_bytes20(input.target);
         return abi.encodePacked(result, bcs_serialize_uint128(input.amount));
     }
@@ -66,11 +53,7 @@ library WrappedFungibleTypes {
         return (new_pos, BurnEvent(target, amount));
     }
 
-    function bcs_deserialize_BurnEvent(bytes memory input)
-        internal
-        pure
-        returns (BurnEvent memory)
-    {
+    function bcs_deserialize_BurnEvent(bytes memory input) internal pure returns (BurnEvent memory) {
         uint256 new_pos;
         BurnEvent memory value;
         (new_pos, value) = bcs_deserialize_offset_BurnEvent(0, input);
@@ -86,29 +69,17 @@ library WrappedFungibleTypes {
         Message_Withdraw withdraw;
     }
 
-    function Message_case_credit(Message_Credit memory credit)
-        internal
-        pure
-        returns (Message memory)
-    {
+    function Message_case_credit(Message_Credit memory credit) internal pure returns (Message memory) {
         Message_Withdraw memory withdraw;
         return Message(uint8(0), credit, withdraw);
     }
 
-    function Message_case_withdraw(Message_Withdraw memory withdraw)
-        internal
-        pure
-        returns (Message memory)
-    {
+    function Message_case_withdraw(Message_Withdraw memory withdraw) internal pure returns (Message memory) {
         Message_Credit memory credit;
         return Message(uint8(1), credit, withdraw);
     }
 
-    function bcs_serialize_Message(Message memory input)
-        internal
-        pure
-        returns (bytes memory)
-    {
+    function bcs_serialize_Message(Message memory input) internal pure returns (bytes memory) {
         if (input.choice == 0) {
             return abi.encodePacked(input.choice, bcs_serialize_Message_Credit(input.credit));
         }
@@ -138,11 +109,7 @@ library WrappedFungibleTypes {
         return (new_pos, Message(choice, credit, withdraw));
     }
 
-    function bcs_deserialize_Message(bytes memory input)
-        internal
-        pure
-        returns (Message memory)
-    {
+    function bcs_deserialize_Message(bytes memory input) internal pure returns (Message memory) {
         uint256 new_pos;
         Message memory value;
         (new_pos, value) = bcs_deserialize_offset_Message(0, input);
@@ -156,11 +123,7 @@ library WrappedFungibleTypes {
         BridgeTypes.AccountOwner source;
     }
 
-    function bcs_serialize_Message_Credit(Message_Credit memory input)
-        internal
-        pure
-        returns (bytes memory)
-    {
+    function bcs_serialize_Message_Credit(Message_Credit memory input) internal pure returns (bytes memory) {
         bytes memory result = BridgeTypes.bcs_serialize_AccountOwner(input.target);
         result = abi.encodePacked(result, bcs_serialize_uint128(input.amount));
         return abi.encodePacked(result, BridgeTypes.bcs_serialize_AccountOwner(input.source));
@@ -181,11 +144,7 @@ library WrappedFungibleTypes {
         return (new_pos, Message_Credit(target, amount, source));
     }
 
-    function bcs_deserialize_Message_Credit(bytes memory input)
-        internal
-        pure
-        returns (Message_Credit memory)
-    {
+    function bcs_deserialize_Message_Credit(bytes memory input) internal pure returns (Message_Credit memory) {
         uint256 new_pos;
         Message_Credit memory value;
         (new_pos, value) = bcs_deserialize_offset_Message_Credit(0, input);
@@ -199,11 +158,7 @@ library WrappedFungibleTypes {
         BridgeTypes.Account target_account;
     }
 
-    function bcs_serialize_Message_Withdraw(Message_Withdraw memory input)
-        internal
-        pure
-        returns (bytes memory)
-    {
+    function bcs_serialize_Message_Withdraw(Message_Withdraw memory input) internal pure returns (bytes memory) {
         bytes memory result = BridgeTypes.bcs_serialize_AccountOwner(input.owner);
         result = abi.encodePacked(result, bcs_serialize_uint128(input.amount));
         return abi.encodePacked(result, BridgeTypes.bcs_serialize_Account(input.target_account));
@@ -224,11 +179,7 @@ library WrappedFungibleTypes {
         return (new_pos, Message_Withdraw(owner, amount, target_account));
     }
 
-    function bcs_deserialize_Message_Withdraw(bytes memory input)
-        internal
-        pure
-        returns (Message_Withdraw memory)
-    {
+    function bcs_deserialize_Message_Withdraw(bytes memory input) internal pure returns (Message_Withdraw memory) {
         uint256 new_pos;
         Message_Withdraw memory value;
         (new_pos, value) = bcs_deserialize_offset_Message_Withdraw(0, input);
@@ -269,14 +220,12 @@ library WrappedFungibleTypes {
         WrappedFungibleOperation_Mint memory mint;
         WrappedFungibleOperation_Burn memory burn;
         WrappedFungibleOperation_RegisterAuthorizedCaller memory register_authorized_caller;
-        return WrappedFungibleOperation(uint8(0), balance_, approve, transfer_, transfer_from, claim, mint, burn, register_authorized_caller);
+        return WrappedFungibleOperation(
+            uint8(0), balance_, approve, transfer_, transfer_from, claim, mint, burn, register_authorized_caller
+        );
     }
 
-    function WrappedFungibleOperation_case_ticker_symbol()
-        internal
-        pure
-        returns (WrappedFungibleOperation memory)
-    {
+    function WrappedFungibleOperation_case_ticker_symbol() internal pure returns (WrappedFungibleOperation memory) {
         WrappedFungibleOperation_Balance memory balance_;
         WrappedFungibleOperation_Approve memory approve;
         WrappedFungibleOperation_Transfer memory transfer_;
@@ -285,7 +234,9 @@ library WrappedFungibleTypes {
         WrappedFungibleOperation_Mint memory mint;
         WrappedFungibleOperation_Burn memory burn;
         WrappedFungibleOperation_RegisterAuthorizedCaller memory register_authorized_caller;
-        return WrappedFungibleOperation(uint8(1), balance_, approve, transfer_, transfer_from, claim, mint, burn, register_authorized_caller);
+        return WrappedFungibleOperation(
+            uint8(1), balance_, approve, transfer_, transfer_from, claim, mint, burn, register_authorized_caller
+        );
     }
 
     function WrappedFungibleOperation_case_approve(WrappedFungibleOperation_Approve memory approve)
@@ -300,7 +251,9 @@ library WrappedFungibleTypes {
         WrappedFungibleOperation_Mint memory mint;
         WrappedFungibleOperation_Burn memory burn;
         WrappedFungibleOperation_RegisterAuthorizedCaller memory register_authorized_caller;
-        return WrappedFungibleOperation(uint8(2), balance_, approve, transfer_, transfer_from, claim, mint, burn, register_authorized_caller);
+        return WrappedFungibleOperation(
+            uint8(2), balance_, approve, transfer_, transfer_from, claim, mint, burn, register_authorized_caller
+        );
     }
 
     function WrappedFungibleOperation_case_transfer(WrappedFungibleOperation_Transfer memory transfer_)
@@ -315,7 +268,9 @@ library WrappedFungibleTypes {
         WrappedFungibleOperation_Mint memory mint;
         WrappedFungibleOperation_Burn memory burn;
         WrappedFungibleOperation_RegisterAuthorizedCaller memory register_authorized_caller;
-        return WrappedFungibleOperation(uint8(3), balance_, approve, transfer_, transfer_from, claim, mint, burn, register_authorized_caller);
+        return WrappedFungibleOperation(
+            uint8(3), balance_, approve, transfer_, transfer_from, claim, mint, burn, register_authorized_caller
+        );
     }
 
     function WrappedFungibleOperation_case_transfer_from(WrappedFungibleOperation_TransferFrom memory transfer_from)
@@ -330,7 +285,9 @@ library WrappedFungibleTypes {
         WrappedFungibleOperation_Mint memory mint;
         WrappedFungibleOperation_Burn memory burn;
         WrappedFungibleOperation_RegisterAuthorizedCaller memory register_authorized_caller;
-        return WrappedFungibleOperation(uint8(4), balance_, approve, transfer_, transfer_from, claim, mint, burn, register_authorized_caller);
+        return WrappedFungibleOperation(
+            uint8(4), balance_, approve, transfer_, transfer_from, claim, mint, burn, register_authorized_caller
+        );
     }
 
     function WrappedFungibleOperation_case_claim(WrappedFungibleOperation_Claim memory claim)
@@ -345,7 +302,9 @@ library WrappedFungibleTypes {
         WrappedFungibleOperation_Mint memory mint;
         WrappedFungibleOperation_Burn memory burn;
         WrappedFungibleOperation_RegisterAuthorizedCaller memory register_authorized_caller;
-        return WrappedFungibleOperation(uint8(5), balance_, approve, transfer_, transfer_from, claim, mint, burn, register_authorized_caller);
+        return WrappedFungibleOperation(
+            uint8(5), balance_, approve, transfer_, transfer_from, claim, mint, burn, register_authorized_caller
+        );
     }
 
     function WrappedFungibleOperation_case_mint(WrappedFungibleOperation_Mint memory mint)
@@ -360,7 +319,9 @@ library WrappedFungibleTypes {
         WrappedFungibleOperation_Claim memory claim;
         WrappedFungibleOperation_Burn memory burn;
         WrappedFungibleOperation_RegisterAuthorizedCaller memory register_authorized_caller;
-        return WrappedFungibleOperation(uint8(6), balance_, approve, transfer_, transfer_from, claim, mint, burn, register_authorized_caller);
+        return WrappedFungibleOperation(
+            uint8(6), balance_, approve, transfer_, transfer_from, claim, mint, burn, register_authorized_caller
+        );
     }
 
     function WrappedFungibleOperation_case_burn(WrappedFungibleOperation_Burn memory burn)
@@ -375,7 +336,9 @@ library WrappedFungibleTypes {
         WrappedFungibleOperation_Claim memory claim;
         WrappedFungibleOperation_Mint memory mint;
         WrappedFungibleOperation_RegisterAuthorizedCaller memory register_authorized_caller;
-        return WrappedFungibleOperation(uint8(7), balance_, approve, transfer_, transfer_from, claim, mint, burn, register_authorized_caller);
+        return WrappedFungibleOperation(
+            uint8(7), balance_, approve, transfer_, transfer_from, claim, mint, burn, register_authorized_caller
+        );
     }
 
     function WrappedFungibleOperation_case_register_authorized_caller(WrappedFungibleOperation_RegisterAuthorizedCaller memory register_authorized_caller)
@@ -390,7 +353,9 @@ library WrappedFungibleTypes {
         WrappedFungibleOperation_Claim memory claim;
         WrappedFungibleOperation_Mint memory mint;
         WrappedFungibleOperation_Burn memory burn;
-        return WrappedFungibleOperation(uint8(8), balance_, approve, transfer_, transfer_from, claim, mint, burn, register_authorized_caller);
+        return WrappedFungibleOperation(
+            uint8(8), balance_, approve, transfer_, transfer_from, claim, mint, burn, register_authorized_caller
+        );
     }
 
     function bcs_serialize_WrappedFungibleOperation(WrappedFungibleOperation memory input)
@@ -408,7 +373,8 @@ library WrappedFungibleTypes {
             return abi.encodePacked(input.choice, bcs_serialize_WrappedFungibleOperation_Transfer(input.transfer_));
         }
         if (input.choice == 4) {
-            return abi.encodePacked(input.choice, bcs_serialize_WrappedFungibleOperation_TransferFrom(input.transfer_from));
+            return
+                abi.encodePacked(input.choice, bcs_serialize_WrappedFungibleOperation_TransferFrom(input.transfer_from));
         }
         if (input.choice == 5) {
             return abi.encodePacked(input.choice, bcs_serialize_WrappedFungibleOperation_Claim(input.claim));
@@ -420,7 +386,10 @@ library WrappedFungibleTypes {
             return abi.encodePacked(input.choice, bcs_serialize_WrappedFungibleOperation_Burn(input.burn));
         }
         if (input.choice == 8) {
-            return abi.encodePacked(input.choice, bcs_serialize_WrappedFungibleOperation_RegisterAuthorizedCaller(input.register_authorized_caller));
+            return abi.encodePacked(
+                input.choice,
+                bcs_serialize_WrappedFungibleOperation_RegisterAuthorizedCaller(input.register_authorized_caller)
+            );
         }
         return abi.encodePacked(input.choice);
     }
@@ -463,10 +432,16 @@ library WrappedFungibleTypes {
         }
         WrappedFungibleOperation_RegisterAuthorizedCaller memory register_authorized_caller;
         if (choice == 8) {
-            (new_pos, register_authorized_caller) = bcs_deserialize_offset_WrappedFungibleOperation_RegisterAuthorizedCaller(new_pos, input);
+            (new_pos, register_authorized_caller) =
+                bcs_deserialize_offset_WrappedFungibleOperation_RegisterAuthorizedCaller(new_pos, input);
         }
         require(choice < 9);
-        return (new_pos, WrappedFungibleOperation(choice, balance_, approve, transfer_, transfer_from, claim, mint, burn, register_authorized_caller));
+        return (
+            new_pos,
+            WrappedFungibleOperation(
+                choice, balance_, approve, transfer_, transfer_from, claim, mint, burn, register_authorized_caller
+            )
+        );
     }
 
     function bcs_deserialize_WrappedFungibleOperation(bytes memory input)
@@ -805,19 +780,11 @@ library WrappedFungibleTypes {
         return value;
     }
 
-    function bcs_serialize_bytes20(bytes20 input)
-        internal
-        pure
-        returns (bytes memory)
-    {
+    function bcs_serialize_bytes20(bytes20 input) internal pure returns (bytes memory) {
         return abi.encodePacked(input);
     }
 
-    function bcs_deserialize_offset_bytes20(uint256 pos, bytes memory input)
-        internal
-        pure
-        returns (uint256, bytes20)
-    {
+    function bcs_deserialize_offset_bytes20(uint256 pos, bytes memory input) internal pure returns (uint256, bytes20) {
         bytes20 dest;
         assembly {
             dest := mload(add(add(input, 0x20), pos))
@@ -825,39 +792,27 @@ library WrappedFungibleTypes {
         return (pos + 20, dest);
     }
 
-    function bcs_serialize_uint128(uint128 input)
-        internal
-        pure
-        returns (bytes memory)
-    {
+    function bcs_serialize_uint128(uint128 input) internal pure returns (bytes memory) {
         bytes memory result = new bytes(16);
         uint128 value = input;
         result[0] = bytes1(uint8(value));
-        for (uint i=1; i<16; i++) {
+        for (uint256 i = 1; i < 16; i++) {
             value = value >> 8;
             result[i] = bytes1(uint8(value));
         }
         return result;
     }
 
-    function bcs_deserialize_offset_uint128(uint256 pos, bytes memory input)
-        internal
-        pure
-        returns (uint256, uint128)
-    {
+    function bcs_deserialize_offset_uint128(uint256 pos, bytes memory input) internal pure returns (uint256, uint128) {
         uint128 value = uint8(input[pos + 15]);
-        for (uint256 i=0; i<15; i++) {
+        for (uint256 i = 0; i < 15; i++) {
             value = value << 8;
             value += uint8(input[pos + 14 - i]);
         }
         return (pos + 16, value);
     }
 
-    function bcs_deserialize_uint128(bytes memory input)
-        internal
-        pure
-        returns (uint128)
-    {
+    function bcs_deserialize_uint128(bytes memory input) internal pure returns (uint128) {
         uint256 new_pos;
         uint128 value;
         (new_pos, value) = bcs_deserialize_offset_uint128(0, input);
@@ -865,33 +820,20 @@ library WrappedFungibleTypes {
         return value;
     }
 
-    function bcs_serialize_uint8(uint8 input)
-        internal
-        pure
-        returns (bytes memory)
-    {
-      return abi.encodePacked(input);
+    function bcs_serialize_uint8(uint8 input) internal pure returns (bytes memory) {
+        return abi.encodePacked(input);
     }
 
-    function bcs_deserialize_offset_uint8(uint256 pos, bytes memory input)
-        internal
-        pure
-        returns (uint256, uint8)
-    {
+    function bcs_deserialize_offset_uint8(uint256 pos, bytes memory input) internal pure returns (uint256, uint8) {
         uint8 value = uint8(input[pos]);
         return (pos + 1, value);
     }
 
-    function bcs_deserialize_uint8(bytes memory input)
-        internal
-        pure
-        returns (uint8)
-    {
+    function bcs_deserialize_uint8(bytes memory input) internal pure returns (uint8) {
         uint256 new_pos;
         uint8 value;
         (new_pos, value) = bcs_deserialize_offset_uint8(0, input);
         require(new_pos == input.length, "incomplete deserialization");
         return value;
     }
-
 } // end of library WrappedFungibleTypes
